Reworked History Tab for KissMP 0.4.5
