+++ kissmp-bridge/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kissmp-bridge"
-version = "0.4.2"
+version = "0.4.5"
 authors = ["hellbox"]
 edition = "2018"
 
@@ -10,6 +10,7 @@ default = ["discord-rpc-client"]
 
 [dependencies]
 shared = { path = "../shared" }
+kissmp-server = { path = "../kissmp-server" }
 bincode = "1.3.2"
 serde = { version = "1.0", features = ["derive"] }
 serde_json="1.0"
