+++ kissmp-master/Cargo.toml
@@ -12,6 +12,6 @@ shared = { path = "../shared" }
 warp = "0.2"
 serde_json="1.0"
 censor = "0.1.1"
-tokio={version = "0.2", features = ["rt-threaded", "macros"]}
+tokio={version = "0.2", features = ["rt-threaded", "macros", "net"]}
 #async-std = { version = "1.8.0", features = ["attributes"] }
 serde = { version = "1.0", features = ["derive"] }
